S3 destination: Extract the same logic for Avro/Parquet formats to separate parent class
