+++ airbyte-integrations/connectors/destination-s3/src/test-integration/java/io/airbyte/integrations/destination/s3/S3AvroDestinationAcceptanceTest.java
@@ -32,8 +32,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
-import org.apache.avro.Schema;
 import org.apache.avro.Schema.Type;
 import org.apache.avro.file.DataFileReader;
 import org.apache.avro.file.SeekableByteArrayInput;
@@ -42,7 +40,7 @@ import org.apache.avro.generic.GenericData.Record;
 import org.apache.avro.generic.GenericDatumReader;
 import org.junit.jupiter.api.Test;
 
-public class S3AvroDestinationAcceptanceTest extends S3DestinationAcceptanceTest {
+public class S3AvroDestinationAcceptanceTest extends S3AvroParquetDestinationAcceptanceTest {
 
   protected S3AvroDestinationAcceptanceTest() {
     super(S3Format.AVRO);
@@ -93,58 +91,8 @@ public class S3AvroDestinationAcceptanceTest extends S3DestinationAcceptanceTest
     return new S3AvroParquetTestDataComparator();
   }
 
-  @Test
-  public void testNumberDataType() throws Exception {
-    final AirbyteCatalog catalog = readCatalogFromFile("number_data_type_test_catalog.json");
-    final ConfiguredAirbyteCatalog configuredCatalog = CatalogHelpers.toDefaultConfiguredCatalog(catalog);
-    final List<AirbyteMessage> messages = readMessagesFromFile("number_data_type_test_messages.txt");
-
-    final JsonNode config = getConfig();
-    final String defaultSchema = getDefaultSchema(config);
-    runSyncAndVerifyStateOutput(config, messages, configuredCatalog, false);
-
-    for (final AirbyteStream stream : catalog.getStreams()) {
-      final String streamName = stream.getName();
-      final String schema = stream.getNamespace() != null ? stream.getNamespace() : defaultSchema;
-
-      Set<Type> actualSchemaTypes = retrieveDataTypesFromSchema(streamName, schema);
-      Optional<Type> actualSchemaTypesWithoutNull = actualSchemaTypes.stream().filter(type -> !type.equals(Type.NULL)).findAny();
-
-      List<Type> expectedTypeList = createSchemaTypesForStreamName(stream.getJsonSchema().get("properties").get("data"));
-      assertEquals(1, expectedTypeList.size(), "Several non null data types are not supported for single stream");
-      assertTrue(actualSchemaTypesWithoutNull.isPresent());
-      assertEquals(expectedTypeList.get(0), actualSchemaTypesWithoutNull.get());
-    }
-  }
-
-  private List<Type> createSchemaTypesForStreamName(JsonNode fieldDefinition) {
-    final JsonNode typeProperty = fieldDefinition.get("type");
-    final JsonNode airbyteTypeProperty = fieldDefinition.get("airbyte_type");
-    final String airbyteTypePropertyText = airbyteTypeProperty == null ? null : airbyteTypeProperty.asText();
-    return Arrays.stream(JsonSchemaType.values())
-        .filter(
-            value -> value.getJsonSchemaType().equals(typeProperty.asText()) && compareAirbyteTypes(airbyteTypePropertyText, value))
-        .map(JsonSchemaType::getAvroType)
-        .toList();
-  }
-
-  private boolean compareAirbyteTypes(String airbyteTypePropertyText, JsonSchemaType value) {
-    if (airbyteTypePropertyText == null){
-      return value.getJsonSchemaAirbyteType() == null;
-    }
-    return airbyteTypePropertyText.equals(value.getJsonSchemaAirbyteType());
-  }
-
-  private AirbyteCatalog readCatalogFromFile(final String catalogFilename) throws IOException {
-    return Jsons.deserialize(MoreResources.readResource(catalogFilename), AirbyteCatalog.class);
-  }
-
-  private List<AirbyteMessage> readMessagesFromFile(final String messagesFilename) throws IOException {
-    return MoreResources.readResource(messagesFilename).lines()
-        .map(record -> Jsons.deserialize(record, AirbyteMessage.class)).collect(Collectors.toList());
-  }
-
-  private Set<Type> retrieveDataTypesFromSchema(final String streamName, final String namespace) throws Exception {
+  @Override
+  protected Set<Type> retrieveDataTypesFromPersistedFiles(final String streamName, final String namespace) throws Exception {
 
     final List<S3ObjectSummary> objectSummaries = getAllSyncedObjects(streamName, namespace);
     Set<Type> dataTypes = new HashSet<>();
@@ -156,22 +104,11 @@ public class S3AvroDestinationAcceptanceTest extends S3DestinationAcceptanceTest
           new GenericDatumReader<>())) {
         while (dataFileReader.hasNext()) {
           final GenericData.Record record = dataFileReader.next();
-          record.getSchema().getField("data").schema();
-          List<Schema> listAvroTypes = record
-              .getSchema()
-              .getField("data")
-              .schema()
-              .getTypes();
-
-          Set<Type> actualDataTypes = listAvroTypes
-              .stream()
-              .map(Schema::getType)
-              .collect(Collectors.toSet());
+          Set<Type> actualDataTypes = getTypes(record);
           dataTypes.addAll(actualDataTypes);
         }
       }
     }
-
     return dataTypes;
   }
